api/apps.py
"""API app configuration."""
from django.apps import AppConfig


class ApiConfig(AppConfig):
    """Configuration for the API app."""
    default_auto_field = 'django.db.models.BigAutoField'
    name = 'api'
    verbose_name = 'Data Extraction API'

